moved method to util
